Works, I hope. I'm not whitelisted -> no e2e
Modifies existing code to make use of React and useEffect to really
drive the wallet interaction. It's crude and entirely out of line with
teh quality of the visuals here.
I think the functions in teh index.js file have been ported to all be in
Hero.js (sorry) where they can get instantiated values instead of having
lots of problems where the javascript would miss - that's the root cause
from what i could tell, for me, on localhost, with the other code.
100% this needs to be tested to make sure I haven't left a glaring hole
in code - but it is the functions from index.js